% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lgbm_predict}
\alias{lgbm_predict}
\title{Predict values using a trained model and recipe}
\usage{
lgbm_predict(spec, recipe, data)
}
\arguments{
\item{spec}{A parsnip model specification object. Must be trained.}

\item{recipe}{A prepped recipe object. Must be trained.}

\item{data}{New data to get predictions from. Will be pre-processed by the
specified \code{recipe}.}
}
\value{
A vector of predictions from the model given the data and recipe
  specified.
}
\description{
Simple helper function to return predictions from a new data set
given a parsnip specification and recipe. Will exponentiate predictions by
default.
}
