% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lgbm_update_params}
\alias{lgbm_update_params}
\title{Finalize workflow objects for LightGBM models}
\usage{
lgbm_update_params(wflow, params)
}
\arguments{
\item{wflow}{A workflow object created by \code{\link[workflows]{workflow}}.}

\item{params}{A dataframe or named list of parameters. Usually created by
\code{\link[tune]{select_best}}.}
}
\value{
A finalized workflow object with updated parameters.
}
\description{
LightGBM is currently not natively supported by Tidymodels.
As such, functions such as \code{\link[tune]{finalize_workflow}}, which is
usually responsible for passing final hyperparameters from cross-validation
into a workflow object, do not function correctly.

This function is a hacky workaround which inserts parameters directly into
the workflow object by looking for their names (workflows are just nested
lists), then inserting the matching values passed in \code{params}.

This function will likely be deprecated by in the near future.
}
