% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{param_lgbm}
\alias{param_lgbm}
\alias{max_bin}
\alias{max_depth}
\alias{add_to_linked_depth}
\alias{learning_rate}
\alias{feature_fraction}
\alias{feature_fraction_bynode}
\alias{min_data_in_leaf}
\alias{min_gain_to_split}
\alias{num_leaves}
\alias{bagging_fraction}
\alias{bagging_freq}
\alias{min_data_per_group}
\alias{max_cat_threshold}
\alias{lambda_l1}
\alias{lambda_l2}
\alias{cat_smooth}
\alias{cat_l2}
\title{LightGBM engine-specific hyperparameters}
\usage{
max_bin(range = c(100L, 500L), trans = NULL)

max_depth(range = c(3L, 17L), trans = NULL)

add_to_linked_depth(range = c(1L, 3L), trans = NULL)

learning_rate(range = c(-3, -0.5), trans = scales::log10_trans())

feature_fraction(range = c(0.3, 1), trans = NULL)

feature_fraction_bynode(range = c(0.3, 1), trans = NULL)

min_data_in_leaf(range = c(20L, 500L), trans = NULL)

min_gain_to_split(range = c(-3, 2), trans = scales::log10_trans())

num_leaves(range = c(31L, 4000L), trans = NULL)

bagging_fraction(range = c(0.3, 1), trans = NULL)

bagging_freq(range = c(1L, 100L), trans = NULL)

min_data_per_group(range = c(20L, 250L), trans = NULL)

max_cat_threshold(range = c(20L, 250L), trans = NULL)

lambda_l1(range = c(-3, 2), trans = scales::log10_trans())

lambda_l2(range = c(-3, 2), trans = scales::log10_trans())

cat_smooth(range = c(10, 100), trans = NULL)

cat_l2(range = c(-3, 2), trans = scales::log10_trans())
}
\arguments{
\item{range}{A two-element vector holding the defaults for the smallest and
largest possible values, respectively.}

\item{trans}{A trans object from the \code{scales} package, such
as \code{\link[scales]{log10_trans}} or
\code{\link[scales]{reciprocal_trans}}. If not provided, the default is
used which matches the units used in range. If no transformation, NULL.}
}
\description{
Collection of hyperparameters specific to LightGBM. Each can be
passed to \code{\link[parsnip]{set_engine}} for tuning. See the LightGBM
\href{https://lightgbm.readthedocs.io/en/latest/Parameters.html}{docs} for
more information on each parameter.
}
\examples{
max_depth()
learning_rate()
feature_fraction()
lambda_l1()
lambda_l2()
cat_smooth()
}
