% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{axe_tune_data}
\alias{axe_tune_data}
\title{Strip attribute data from CV objects to shrink their size}
\usage{
axe_tune_data(x)
}
\arguments{
\item{x}{Iteration object created by the tune package.}
}
\value{
Iteration object from \code{x} but with CV split and attributes
  removed.
}
\description{
Remove unnecessary attribute data from iteration results objects
created by \code{\link[tune]{tune_grid}} and \code{\link[tune]{tune_bayes}}.
This makes the parameter object significantly smaller when saved to disk.
}
\seealso{
Other model_funs: 
\code{\link{axe_recipe}()},
\code{\link{lgbm_load}()},
\code{\link{lgbm_save}()},
\code{\link{lgbm_update_params}()}
}
\concept{model_funs}
