% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.R
\name{train_lightgbm}
\alias{train_lightgbm}
\title{Boosted trees via LightGBM}
\usage{
train_lightgbm(
  x,
  y,
  num_iterations = 10,
  max_depth = 17,
  num_leaves = 31,
  learning_rate = 0.1,
  feature_fraction = 1,
  min_data_in_leaf = 20,
  min_gain_to_split = 0,
  link_max_depth = FALSE,
  add_to_linked_depth = 2L,
  categorical_feature = NULL,
  max_bin = NULL,
  feature_pre_filter = FALSE,
  verbose = 0,
  ...
)
}
\arguments{
\item{x}{A matrix of predictors.}

\item{y}{A numeric vector of outcome data.}

\item{num_iterations}{Integer value for the number of iterations (trees)
to grow.}

\item{max_depth}{Integer value for the maximum leaf distance from the root
node.}

\item{num_leaves}{Integer value for the maximum possible number of leaves
in one tree.}

\item{learning_rate}{A numeric value between zero and one to control
the learning rate.}

\item{feature_fraction}{Subsampling proportion of columns.}

\item{min_data_in_leaf}{A numeric value for the minimum sum of instances
needed in a child to continue to split.}

\item{min_gain_to_split}{A number for the minimum loss reduction required
to make a further partition on a leaf node of the tree.}

\item{link_max_depth}{Logical, default FALSE. When TRUE, and when
\code{max_depth} is unconstrained \code{-1}, then \code{max_depth} will
be set to \code{floor(log2(num_leaves)) + link_max_depth_add}.}

\item{add_to_linked_depth}{Integer value to add to \code{max_depth} when it
is linked to \code{num_leaves}.}

\item{categorical_feature}{A character vector of feature names or an
integer vector with the indices of the features.}

\item{max_bin}{Max number of bins that feature values will be bucketed in.}

\item{feature_pre_filter}{Tell LightGBM to ignore the features that are
unsplittable based on \code{min_data_in_leaf}.}

\item{verbose}{Integer. < 0: Fatal, = 0: Error (Warning), = 1: Info,
> 1: Debug.}

\item{...}{Engine arguments, hyperparameters, etc. that are passed on to
\code{\link[lightgbm]{lgb.train}}.}
}
\value{
A fitted \code{lgb.Booster} object.
}
\description{
Wrapper for \code{\link[lightgbm]{lgb.train}} tree-based models
with some expanded/advanced options.
}
